LayoutTests/fast/profiler/nested-start-and-stop-profiler.html

<html>
<head>
<script src="resources/profiler-test-JS-resources.js"></script>
<script>
if (window.testRunner)
    testRunner.dumpAsText();
if (window.internals)
    internals.setJavaScriptProfilingEnabled(true);

function startTest()
{
    console.profile("Start the profiler the first time.");
    console.profile("Start the profiler the second time.");
    console.profile("Start the profiler the third time.");
    endTest();
    endTest();
    endTest();
}
</script>
</head>

<body onload="startTest()">
This page's JavaScript calls console.profile() three times then console.profileEnd()
three times.
<br>
<br>
To run this test manually, load it in the browser then load the WebInspector and look at
the profile.  In the profiler there should be three profiles.
<div id="output"></div>
</body>
</html>
